refactoring, WIP, specs not passing
