+++ services/QuillLMS/engines/evidence/app/controllers/evidence/opinion_controller.rb
@@ -1,19 +1,9 @@
 module Evidence
   class OpinionController < ApiController
-    API_TIMEOUT = 5
-    def fetch
-      oapi_response = Timeout.timeout(API_TIMEOUT) do 
-        HTTParty.post(
-          ENV['OPINION_API_DOMAIN'], 
-          headers:  {'Content-Type': 'application/json'},
-          body:     {
-            entry: params['entry'],
-            prompt_text: params['prompt_text']
-          }.to_json
-        )
-      end
 
-      assembler = OpinionFeedbackAssembler.new(oapi_response)
+    def fetch
+      oapi_client = Opinion::Client.new(entry: params['entry'], prompt_text: params['prompt_text'])
+      assembler = Opinion::FeedbackAssembler.new(oapi_client.post)
       render json: assembler.to_payload
     end
   end
